+++ rotkehlchen/chain/aggregator.py
@@ -154,6 +154,8 @@ DEFI_PROTOCOLS_TO_SKIP_ASSETS = {
     # aTokens are already detected at token balance queries
     'Aave': True,  # True means all
     'Aave V2': True,  # True means all
+    # stkAAVE and staking incentives are already detected
+    'Aave • Staking': True,
     # cTokens are already detected at token balance queries
     'Compound': True,  # True means all
     # Curve balances are detected by our scan for ERC20 tokens